requirements.txt
numpy
pandas
python-dotenv
psycopg2
psycopg2
scikit-learn

#-e .